requirements.txt
pywebview==4.2.2
pywin32==306
requests==2.31.0